+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.9] - 21-09-2021
+
+### Changed
+- Inputs on tab fixed
+- Rule enabled toggle bugfix
+- Removed JsonStringEnumConverter until WASM PWA supported
+- Workflow/Input toggler now based on DIV hidden toggle
+
+### Added
+- InputRule is now InputRuleName (former deprecated as Obsolete until next ver)
+- InputRule.Parameter is now Parameters (former deprecated as Obsolete until next ver)
+- Parameter Add assist
+
 ## [1.0.0] - 14-09-2021
 
 ### Added